feat: basic functionality for memory mapped IO

+++ src/device/mod.rs
@@ -6,7 +6,6 @@ pub use crate::device::rom::Rom;
 
 /// Trait for memory-mapped I/O devices.
 pub trait MMIODevice {
-    fn read(&self, address: usize) -> usize;
-    fn write(&mut self, address: usize, value: usize);
+    fn read(&self, address: usize) -> u32;
+    fn write(&mut self, address: usize, value: u32);
 }
-
